aa/backend/ai_rag_service/requirements.txt
Flask==3.1.0
Flask-Cors==5.0.0
requests==2.32.3
python-dotenv==0.19.1

supabase==2.15.1

# For JWT support
pyjwt==2.6.0

openai==1.76.0
pydantic==2.11.1